Time: 0 ms (100%), Space: 11 MB (48.42%) - LeetHub

+++ 2436-make-array-zero-by-subtracting-equal-amounts/2436-make-array-zero-by-subtracting-equal-amounts.cpp
@@ -10,6 +10,6 @@ public:
             }
         }
 
-        return uniqueValsMap.size();
+        return static_cast<int>(uniqueValsMap.size());
     }
 };
